Fix bug when receive data buffer has multiple packets
The problem is that currently we block until all data in the snoop socket is received, potentially including multiple HCI packets. Then we parse and handle the first packet only and continue to block until more data is available in the socket. This means that each packet is parsed after an unexpected delay, which is the time it takes for new data to be available in the socket.
This caused the HCI packet handling methods to assume that the controller is unresponsive.

When the buffer contains more than one buffer, we should continue handling all remaining packets before blocking until the snoop socket has more data.